// Frontend/front-university/src/app/services/api/models/register-student.ts
/* tslint:disable */
/* eslint-disable */
export interface RegisterStudent {
  id: number;
  city?: null | string;
  comunity?: null | string;
  country?: null | string;
  dob: Date;
  email: string;
  firstName: string;
  lastName: string;
  password: string;
  state?: null | string;
  street?: null | string;
  zipCode?: null | string;
}